Remove debugging print statement
This was not a bug, just a "feature" of the defaults in Terra. The
workflow was succeeding because call caching was on by default when
submitting the job, so the workflow ran and got to the step of
checking whether the directory already existed. It just copied the
old output.